test_cases/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==1.3.0
attrs==22.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
core-universal==4.0.2
cryptography==41.0.7
execnet==2.0.2
eyes-common==5.22.2
eyes-selenium==5.21.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
marshmallow==2.21.0
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
pytest==7.4.2
pytest-xdist==3.4.0
python-dotenv==1.0.0
requests==2.31.0
selenium==3.141.0
six==1.16.0
smart-assertions==1.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
websocket-client==0.59.0
wsproto==1.2.0
